completed hardware_service config
