+++ conf/quads.yml
@@ -10,7 +10,7 @@ log: /opt/quads/log/quads.log
 
 # EC528 addition - demo 3
 # set hardware service - currently 3 options: MockDriver, HilDriver, and JuniperDriver
-hardware_service: MockDriver
+hardware_service: JuniperDriver
