+++ lib/Quads.py
@@ -19,14 +19,15 @@ import yaml
 import os
 import logging
 import requests
-from subprocess import call
+import sys
+import importlib
 from Clouds import Clouds
 from History import History
 from QuadsData import QuadsData
 from CloudHistory import CloudHistory
 from hardware_services.hardware_service import get_hardware_service, set_hardware_service
 from hardware_services.hardware_drivers.juniper_driver import JuniperDriver
-from hardware_services.hardware_drivers.hil_driver import HilDriver #HIL Driver importing
+from hardware_services.hardware_drivers.hil_driver import HilDriver
 
 
 class Quads(object):
@@ -41,9 +42,11 @@ class Quads(object):
         self.datearg = datearg
         self.logger = logging.getLogger("quads.Quads")
         self.logger.setLevel(logging.DEBUG)
-        #set_hardware_service(JuniperDriver())
-	set_hardware_service(HilDriver())  #HIL driver addition
 
+        #EC528 addition - dynamically import driver module and set hardware service
+        importlib.import_module(hardwareservice)
+        set_hardware_service(getattr(sys.modules[hardwareservice], hardwareservice)())
+        self.hardware_service = get_hardware_service()
 
         if initialize:
             self.quads_init_data(force)
@@ -219,12 +222,12 @@ class Quads(object):
     # list the hosts
     def quads_list_hosts(self):
         # list just the hostnames
-        get_hardware_service().list_hosts(self)
+        self.hardware_service.list_hosts(self)
 
     # list the hosts
     def quads_list_clouds(self):
         # list just the hostnames
-        get_hardware_service().list_clouds(self)
+        self.hardware_service.list_clouds(self)
 
     # list the owners
     def quads_list_owners(self, cloudonly):
@@ -296,7 +299,7 @@ class Quads(object):
 
         kwargs = {'rmhost': rmhost}
 
-        get_hardware_service().remove_host(self, **kwargs)
+        self.hardware_service.remove_host(self, **kwargs)
 
         return
 
@@ -306,7 +309,7 @@ class Quads(object):
 
         kwargs = {'rmcloud': rmcloud}
 
-        get_hardware_service().remove_cloud(self, **kwargs)
+        self.hardware_service.remove_cloud(self, **kwargs)
 
         return
 
@@ -316,7 +319,7 @@ class Quads(object):
 
         kwargs = {'hostresource': hostresource, 'hostcloud': hostcloud, 'forceupdate': forceupdate}
 
-        get_hardware_service().update_host(self, **kwargs)
+        self.hardware_service.update_host(self, **kwargs)
 
         return
 
@@ -327,7 +330,7 @@ class Quads(object):
         kwargs = {'cloudresource': cloudresource, 'description': description, 'forceupdate': forceupdate,
                   'cloudowner': cloudowner, 'ccusers': ccusers, 'cloudticket': cloudticket, 'qinq': qinq}
 
-        get_hardware_service().update_cloud(self, **kwargs)
+        self.hardware_service.update_cloud(self, **kwargs)
 
         return
 
@@ -511,7 +514,7 @@ class Quads(object):
         kwargs = {'movecommand': movecommand, 'dryrun': dryrun, 'statedir': statedir,
                   'datearg': datearg}
 
-        get_hardware_service().move_hosts(self, **kwargs)
+        self.hardware_service.move_hosts(self, **kwargs)
 
         exit(0)
